Deleted some unused files, and added one thai unicode to tis converter file. Makefile changes for abi required again I suppose.  In groundbreaking news two microsoft insanities identified and removed.
Lists are really crazy. Lists information is stored seperately from the entries
involved, and indexed in the pap. I have come across negative values for this!,
normalizing them to poistive appears to do the right thing. Also I have reworked
the "invalid list entry" code a bit, this is where the format promises that a
number of list overrides are forthcoming in the file, and then inserts
0xffffffff as some sort of marker. Maybe it means that "there are no list
overrides after all", but I believe that it just means "ignore me and try
again after me". Following that rule creates the possibility of heading off
the end of the file, so a test for that condition is also included. This wipes
out a few more broken files.

Tables have a table look specifier which sets what way the table should be
colored in, and what sort of borders it should have. There are 40 table
looks that can be selected. Needless to say there are files that index a
value way above 40, this also has to be caught and munged into a valid
number.

Also I have added a few sanity checks here and there, primarily just so
that I could get one file to convert that was crashing MSWord97 itself. So
that I could feel smug about it all :-)

C.


git-svn-id: http://svn.abisource.com/wv/trunk@20331 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6